Integrate real chat and certificate data with backend
Replaced mock data with real API calls for chat and certificate features. ChatComponente now loads and sends messages via the backend, with polling and user context. ChatsContext fetches active orders as chat threads. CertificadoPagina and DescargaPagina now fetch and map certificate data from the backend, including environmental metrics. SecurityConfig updated to allow chat endpoints and clarify authority checks. CertificadoController endpoint updated to accept Integer for ordenId. UI improvements for chat and certificate flows.

+++ src/main/java/com/greencycle/ecoswap/ecoswap/controller/CertificadoController.java
@@ -14,11 +14,10 @@ public class CertificadoController {
     @Autowired
     private CertificadoService certificadoService;
 
-    // Obtener los datos calculados de una orden finalizada
     @GetMapping("/{ordenId}")
-    public ResponseEntity<CertificadoResponse> obtenerCertificado(@PathVariable Long ordenId) {
+    public ResponseEntity<CertificadoResponse> obtenerCertificado(@PathVariable Integer ordenId) {
         try {
-            CertificadoResponse response = certificadoService.obtenerDatosCertificado(Math.toIntExact(ordenId));
+            CertificadoResponse response = certificadoService.obtenerDatosCertificado(ordenId);
             return ResponseEntity.ok(response);
         } catch (RuntimeException e) {
             return ResponseEntity.notFound().build();
